Added command line functionality
Added the ability to use the basic functionality of our project
from the command line. This includes uploading a file, checking
the hash of a file with that of the server, and basic closing.
As of now, the main thing left to do is tackle how we can
parallelize the program. I'll add comments to my code tomorrow.
It should be relatively simple from here. I may need to modify
the printed statements and server interaction to allow the user
the ability to see the name of the file that was sent. This will
also help us when we parallelize our system, allowing us to see
the files being sent and then the information of thise files
being returned by the server in different orders. Nonetheless,
this is the progress so far.